Refactor: axios 함수 파일 수정
